refactor(diff): refactor diff functionality
 * move calculateTotalCosts to output since we don't need it in the schema.
 * remove breakdown struct from schema, since we don't need the total costs here so this simplifies it, reduces nil ptr possibilities
 * move diff logic to separate file
